refactor: Extract executor to break CI import cycle and consolidate token resolution
Break the import cycle (pkg/ci → artifact/github → providers/github → pkg/ci) by
extracting the executor into pkg/ci/executor/ with function-pointer registration.
Consolidate CI GitHub token resolution into a single code path in
pkg/ci/providers/github/token.go so both the provider client and artifact store
use the same detection chain (ATMOS_CI_GITHUB_TOKEN > ATMOS_GITHUB_TOKEN >
GITHUB_TOKEN > gh auth token).
