+++ pkg/ci/executor_test.go
@@ -1,10 +1,6 @@
 package ci
 
 import (
-	"context"
-	"fmt"
-	"os"
-	"path/filepath"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -12,41 +8,7 @@ import (
 
 	plugin "github.com/cloudposse/atmos/pkg/ci/internal/plugin"
 	"github.com/cloudposse/atmos/pkg/ci/internal/provider"
-	"github.com/cloudposse/atmos/pkg/schema"
 )
-
-// boolPtr returns a pointer to a bool value.
-func boolPtr(b bool) *bool {
-	return &b
-}
-
-// capturingProvider is a mock CI provider that records CreateCheckRun and UpdateCheckRun calls.
-type capturingProvider struct {
-	mockProvider
-	createCheckRunCalls []*provider.CreateCheckRunOptions
-	updateCheckRunCalls []*provider.UpdateCheckRunOptions
-	nextID              int64
-}
-
-func (c *capturingProvider) CreateCheckRun(_ context.Context, opts *provider.CreateCheckRunOptions) (*provider.CheckRun, error) {
-	c.createCheckRunCalls = append(c.createCheckRunCalls, opts)
-	c.nextID++
-	return &provider.CheckRun{
-		ID:     c.nextID,
-		Name:   opts.Name,
-		Status: opts.Status,
-	}, nil
-}
-
-func (c *capturingProvider) UpdateCheckRun(_ context.Context, opts *provider.UpdateCheckRunOptions) (*provider.CheckRun, error) {
-	c.updateCheckRunCalls = append(c.updateCheckRunCalls, opts)
-	c.nextID++
-	return &provider.CheckRun{
-		ID:     c.nextID,
-		Name:   opts.Name,
-		Status: opts.Status,
-	}, nil
-}
 
 // stubPlugin is a simple Plugin for tests that uses Handler callbacks.
 type stubPlugin struct {
@@ -60,482 +22,6 @@ func (s *stubPlugin) GetType() string {
 
 func (s *stubPlugin) GetHookBindings() []plugin.HookBinding {
 	return s.bindings
-}
-
-func TestExecute_BeforeTerraformPlan_TriggersCheckCreate(t *testing.T) {
-	// Set up provider registry with a capturing provider.
-	backup := testSaveAndClearRegistry()
-	defer testRestoreRegistry(backup)
-
-	cp := &capturingProvider{
-		mockProvider: mockProvider{name: "generic", detected: false},
-	}
-	Register(cp)
-
-	// Set up plugin with a handler that creates a check run.
-	ClearPlugins()
-	sp := &stubPlugin{
-		componentType: "terraform",
-		bindings: []plugin.HookBinding{
-			{
-				Event: "before.terraform.plan",
-				Handler: func(ctx *plugin.HookContext) error {
-					// Simulate check run creation.
-					name := provider.FormatCheckRunName(ctx.Command, ctx.Info.Stack, ctx.Info.ComponentFromArg)
-					opts := &provider.CreateCheckRunOptions{
-						Name:   name,
-						Status: provider.CheckRunStateInProgress,
-						Title:  fmt.Sprintf("Running %s...", ctx.Command),
-					}
-					if ctx.CICtx != nil {
-						opts.Owner = ctx.CICtx.RepoOwner
-						opts.Repo = ctx.CICtx.RepoName
-						opts.SHA = ctx.CICtx.SHA
-					}
-					_, err := ctx.Provider.CreateCheckRun(context.Background(), opts)
-					return err
-				},
-			},
-		},
-	}
-	err := RegisterPlugin(sp)
-	require.NoError(t, err)
-
-	// Execute with CI forced, checks enabled.
-	err = Execute(ExecuteOptions{
-		Event:       "before.terraform.plan",
-		ForceCIMode: true,
-		AtmosConfig: &schema.AtmosConfiguration{
-			CI: schema.CIConfig{
-				Checks: schema.CIChecksConfig{Enabled: boolPtr(true)},
-			},
-		},
-		Info: &schema.ConfigAndStacksInfo{
-			Stack:            "dev-us-east-1",
-			ComponentFromArg: "vpc",
-		},
-	})
-	require.NoError(t, err)
-
-	// Verify CreateCheckRun was called.
-	require.Len(t, cp.createCheckRunCalls, 1, "CreateCheckRun should have been called once")
-	assert.Equal(t, "atmos/plan: dev-us-east-1/vpc", cp.createCheckRunCalls[0].Name)
-	assert.Equal(t, provider.CheckRunStateInProgress, cp.createCheckRunCalls[0].Status)
-}
-
-func TestExecute_BeforeAfterPlan_CheckLifecycle(t *testing.T) {
-	// Set up provider registry with a capturing provider.
-	backup := testSaveAndClearRegistry()
-	defer testRestoreRegistry(backup)
-
-	cp := &capturingProvider{
-		mockProvider: mockProvider{name: "generic", detected: false},
-	}
-	Register(cp)
-
-	// Set up plugin with before and after plan handlers that manage check runs.
-	ClearPlugins()
-	sp := &stubPlugin{
-		componentType: "terraform",
-		bindings: []plugin.HookBinding{
-			{
-				Event: "before.terraform.plan",
-				Handler: func(ctx *plugin.HookContext) error {
-					name := provider.FormatCheckRunName(ctx.Command, ctx.Info.Stack, ctx.Info.ComponentFromArg)
-					_, err := ctx.Provider.CreateCheckRun(context.Background(), &provider.CreateCheckRunOptions{
-						Name:   name,
-						Status: provider.CheckRunStateInProgress,
-					})
-					return err
-				},
-			},
-			{
-				Event: "after.terraform.plan",
-				Handler: func(ctx *plugin.HookContext) error {
-					name := provider.FormatCheckRunName(ctx.Command, ctx.Info.Stack, ctx.Info.ComponentFromArg)
-					_, err := ctx.Provider.UpdateCheckRun(context.Background(), &provider.UpdateCheckRunOptions{
-						Name:       name,
-						Status:     provider.CheckRunStateSuccess,
-						Conclusion: "success",
-						Title:      "Plan: 1 to add, 0 to change, 0 to destroy.",
-					})
-					return err
-				},
-			},
-		},
-	}
-	err := RegisterPlugin(sp)
-	require.NoError(t, err)
-
-	atmosConfig := &schema.AtmosConfiguration{
-		CI: schema.CIConfig{
-			Checks: schema.CIChecksConfig{Enabled: boolPtr(true)},
-		},
-	}
-	info := &schema.ConfigAndStacksInfo{
-		Stack:            "dev-us-east-1",
-		ComponentFromArg: "vpc",
-	}
-
-	// 1. Execute "before.terraform.plan" — should create a check run.
-	err = Execute(ExecuteOptions{
-		Event:       "before.terraform.plan",
-		ForceCIMode: true,
-		AtmosConfig: atmosConfig,
-		Info:        info,
-	})
-	require.NoError(t, err)
-	require.Len(t, cp.createCheckRunCalls, 1, "CreateCheckRun should have been called once after before event")
-	assert.Equal(t, provider.CheckRunStateInProgress, cp.createCheckRunCalls[0].Status)
-
-	// 2. Execute "after.terraform.plan" — should update the check run, not create a new one.
-	err = Execute(ExecuteOptions{
-		Event:       "after.terraform.plan",
-		ForceCIMode: true,
-		AtmosConfig: atmosConfig,
-		Info:        info,
-		Output:      "Terraform will perform the following actions...",
-	})
-	require.NoError(t, err)
-
-	// Verify UpdateCheckRun was called (not a second CreateCheckRun).
-	require.Len(t, cp.updateCheckRunCalls, 1, "UpdateCheckRun should have been called once after after event")
-	assert.Equal(t, provider.CheckRunStateSuccess, cp.updateCheckRunCalls[0].Status)
-	assert.Equal(t, "success", cp.updateCheckRunCalls[0].Conclusion)
-	assert.Equal(t, "Plan: 1 to add, 0 to change, 0 to destroy.", cp.updateCheckRunCalls[0].Title)
-
-	// CreateCheckRun should NOT have been called again.
-	assert.Len(t, cp.createCheckRunCalls, 1, "CreateCheckRun should not be called during after event when ID is stored")
-}
-
-func TestExecute_CIEnabledInConfig_NoForceFlag_NoActions(t *testing.T) {
-	// When CI.Enabled=true in atmos.yaml but --ci flag is NOT passed and
-	// no CI platform is detected, Execute should be a no-op.
-	backup := testSaveAndClearRegistry()
-	defer testRestoreRegistry(backup)
-
-	cp := &capturingProvider{
-		mockProvider: mockProvider{name: "generic", detected: false},
-	}
-	Register(cp)
-
-	ClearPlugins()
-	handlerCalled := false
-	sp := &stubPlugin{
-		componentType: "terraform",
-		bindings: []plugin.HookBinding{
-			{
-				Event: "before.terraform.plan",
-				Handler: func(_ *plugin.HookContext) error {
-					handlerCalled = true
-					return nil
-				},
-			},
-		},
-	}
-	err := RegisterPlugin(sp)
-	require.NoError(t, err)
-
-	// Execute WITHOUT ForceCIMode — CI is enabled only via config.
-	err = Execute(ExecuteOptions{
-		Event:       "before.terraform.plan",
-		ForceCIMode: false,
-		AtmosConfig: &schema.AtmosConfiguration{
-			CI: schema.CIConfig{
-				Enabled: true,
-				Checks:  schema.CIChecksConfig{Enabled: boolPtr(true)},
-			},
-		},
-		Info: &schema.ConfigAndStacksInfo{
-			Stack:            "dev-us-east-1",
-			ComponentFromArg: "vpc",
-		},
-	})
-	require.NoError(t, err)
-
-	// No CI actions should run — ci.enabled alone does not trigger the generic fallback.
-	assert.False(t, handlerCalled, "Handler should NOT be called without --ci flag even when CI.Enabled=true in config")
-}
-
-// capturingOutputWriter records all WriteSummary and WriteOutput calls.
-type capturingOutputWriter struct {
-	summaries []string
-	outputs   map[string]string
-}
-
-func newCapturingOutputWriter() *capturingOutputWriter {
-	return &capturingOutputWriter{outputs: make(map[string]string)}
-}
-
-func (w *capturingOutputWriter) WriteSummary(content string) error {
-	w.summaries = append(w.summaries, content)
-	return nil
-}
-
-func (w *capturingOutputWriter) WriteOutput(key, value string) error {
-	w.outputs[key] = value
-	return nil
-}
-
-// fullCapturingProvider captures all CI operations: check runs AND output writes.
-type fullCapturingProvider struct {
-	mockProvider
-	createCheckRunCalls []*provider.CreateCheckRunOptions
-	updateCheckRunCalls []*provider.UpdateCheckRunOptions
-	writer              *capturingOutputWriter
-	nextID              int64
-}
-
-func newFullCapturingProvider(name string, detected bool) *fullCapturingProvider {
-	return &fullCapturingProvider{
-		mockProvider: mockProvider{name: name, detected: detected},
-		writer:       newCapturingOutputWriter(),
-	}
-}
-
-func (c *fullCapturingProvider) CreateCheckRun(_ context.Context, opts *provider.CreateCheckRunOptions) (*provider.CheckRun, error) {
-	c.createCheckRunCalls = append(c.createCheckRunCalls, opts)
-	c.nextID++
-	return &provider.CheckRun{ID: c.nextID, Name: opts.Name, Status: opts.Status}, nil
-}
-
-func (c *fullCapturingProvider) UpdateCheckRun(_ context.Context, opts *provider.UpdateCheckRunOptions) (*provider.CheckRun, error) {
-	c.updateCheckRunCalls = append(c.updateCheckRunCalls, opts)
-	c.nextID++
-	return &provider.CheckRun{ID: c.nextID, Name: opts.Name, Status: opts.Status}, nil
-}
-
-func (c *fullCapturingProvider) OutputWriter() provider.OutputWriter {
-	return c.writer
-}
-
-func (c *fullCapturingProvider) totalCICalls() int {
-	return len(c.createCheckRunCalls) + len(c.updateCheckRunCalls) +
-		len(c.writer.summaries) + len(c.writer.outputs)
-}
-
-// TestExecute_LocalDev_NoCIFlag_NoActions verifies that CI functionality
-// (summaries, checks, outputs) is never triggered on a local developer
-// machine when the --ci flag is not set and CI is not enabled in config.
-func TestExecute_LocalDev_NoCIFlag_NoActions(t *testing.T) {
-	events := []string{"before.terraform.plan", "after.terraform.plan"}
-
-	createAllBindings := func() []plugin.HookBinding {
-		handlerCalled := false
-		return []plugin.HookBinding{
-			{
-				Event: "before.terraform.plan",
-				Handler: func(_ *plugin.HookContext) error {
-					handlerCalled = true
-					_ = handlerCalled
-					return nil
-				},
-			},
-			{
-				Event: "after.terraform.plan",
-				Handler: func(_ *plugin.HookContext) error {
-					handlerCalled = true
-					_ = handlerCalled
-					return nil
-				},
-			},
-		}
-	}
-
-	t.Run("no CI flag, no CI config — no actions executed", func(t *testing.T) {
-		backup := testSaveAndClearRegistry()
-		defer testRestoreRegistry(backup)
-
-		cp := newFullCapturingProvider("generic", false)
-		Register(cp)
-
-		ClearPlugins()
-		sp := &stubPlugin{componentType: "terraform", bindings: createAllBindings()}
-		require.NoError(t, RegisterPlugin(sp))
-
-		for _, event := range events {
-			err := Execute(ExecuteOptions{
-				Event:       event,
-				ForceCIMode: false,
-				AtmosConfig: &schema.AtmosConfiguration{},
-				Info: &schema.ConfigAndStacksInfo{
-					Stack:            "dev-us-east-1",
-					ComponentFromArg: "vpc",
-				},
-			})
-			assert.NoError(t, err)
-		}
-
-		assert.Zero(t, cp.totalCICalls(),
-			"No CI actions should execute on local dev without --ci flag")
-	})
-
-	t.Run("no CI flag, CI.Enabled explicitly false — no actions executed", func(t *testing.T) {
-		backup := testSaveAndClearRegistry()
-		defer testRestoreRegistry(backup)
-
-		cp := newFullCapturingProvider("generic", false)
-		Register(cp)
-
-		ClearPlugins()
-		sp := &stubPlugin{componentType: "terraform", bindings: createAllBindings()}
-		require.NoError(t, RegisterPlugin(sp))
-
-		for _, event := range events {
-			err := Execute(ExecuteOptions{
-				Event:       event,
-				ForceCIMode: false,
-				AtmosConfig: &schema.AtmosConfiguration{
-					CI: schema.CIConfig{Enabled: false},
-				},
-				Info: &schema.ConfigAndStacksInfo{
-					Stack:            "dev-us-east-1",
-					ComponentFromArg: "vpc",
-				},
-			})
-			assert.NoError(t, err)
-		}
-
-		assert.Zero(t, cp.totalCICalls(),
-			"No CI actions should execute when CI is explicitly disabled")
-	})
-
-	t.Run("no CI flag, nil AtmosConfig — no actions executed", func(t *testing.T) {
-		backup := testSaveAndClearRegistry()
-		defer testRestoreRegistry(backup)
-
-		cp := newFullCapturingProvider("generic", false)
-		Register(cp)
-
-		ClearPlugins()
-		sp := &stubPlugin{componentType: "terraform", bindings: createAllBindings()}
-		require.NoError(t, RegisterPlugin(sp))
-
-		for _, event := range events {
-			err := Execute(ExecuteOptions{
-				Event:       event,
-				ForceCIMode: false,
-				AtmosConfig: nil,
-				Info: &schema.ConfigAndStacksInfo{
-					Stack:            "dev-us-east-1",
-					ComponentFromArg: "vpc",
-				},
-			})
-			assert.NoError(t, err)
-		}
-
-		assert.Zero(t, cp.totalCICalls(),
-			"No CI actions should execute when AtmosConfig is nil")
-	})
-
-	t.Run("no CI flag, checks enabled but CI.Enabled false — no actions executed", func(t *testing.T) {
-		backup := testSaveAndClearRegistry()
-		defer testRestoreRegistry(backup)
-
-		cp := newFullCapturingProvider("generic", false)
-		Register(cp)
-
-		ClearPlugins()
-		sp := &stubPlugin{componentType: "terraform", bindings: createAllBindings()}
-		require.NoError(t, RegisterPlugin(sp))
-
-		for _, event := range events {
-			err := Execute(ExecuteOptions{
-				Event:       event,
-				ForceCIMode: false,
-				AtmosConfig: &schema.AtmosConfiguration{
-					CI: schema.CIConfig{
-						Enabled: false,
-						Checks:  schema.CIChecksConfig{Enabled: boolPtr(true)},
-						Summary: schema.CISummaryConfig{Enabled: boolPtr(true)},
-						Output:  schema.CIOutputConfig{Enabled: boolPtr(true)},
-					},
-				},
-				Info: &schema.ConfigAndStacksInfo{
-					Stack:            "dev-us-east-1",
-					ComponentFromArg: "vpc",
-				},
-			})
-			assert.NoError(t, err)
-		}
-
-		assert.Zero(t, cp.totalCICalls(),
-			"No CI actions should execute when CI.Enabled is false, even if individual actions are enabled")
-	})
-}
-
-func TestExecute(t *testing.T) {
-	t.Run("returns nil when platform not detected and force mode disabled", func(t *testing.T) {
-		backup := testSaveAndClearRegistry()
-		defer testRestoreRegistry(backup)
-
-		err := Execute(ExecuteOptions{
-			Event:       "after.terraform.plan",
-			ForceCIMode: false,
-		})
-		assert.NoError(t, err)
-	})
-
-	t.Run("uses generic provider when force mode enabled and no platform detected", func(t *testing.T) {
-		backup := testSaveAndClearRegistry()
-		defer testRestoreRegistry(backup)
-		// Register a mock "generic" provider so Get("generic") succeeds.
-		Register(&mockProvider{name: "generic", detected: false})
-
-		// Force CI mode should use generic provider.
-		err := Execute(ExecuteOptions{
-			Event:       "after.terraform.plan",
-			ForceCIMode: true,
-			Info: &schema.ConfigAndStacksInfo{
-				ComponentFromArg: "vpc",
-				Stack:            "dev",
-			},
-		})
-		// Should not error - generic provider will handle it.
-		assert.NoError(t, err)
-	})
-}
-
-func TestExtractComponentType(t *testing.T) {
-	tests := []struct {
-		event    string
-		expected string
-	}{
-		{"after.terraform.plan", "terraform"},
-		{"before.terraform.apply", "terraform"},
-		{"after.helmfile.diff", "helmfile"},
-		{"invalid", ""},
-		{"single", ""},
-		{"", ""},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.event, func(t *testing.T) {
-			result := extractComponentType(tt.event)
-			assert.Equal(t, tt.expected, result)
-		})
-	}
-}
-
-func TestExtractCommand(t *testing.T) {
-	tests := []struct {
-		event    string
-		expected string
-	}{
-		{"after.terraform.plan", "plan"},
-		{"before.terraform.apply", "apply"},
-		{"after.helmfile.diff", "diff"},
-		{"invalid", ""},
-		{"single.part", ""},
-		{"", ""},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.event, func(t *testing.T) {
-			result := extractCommand(tt.event)
-			assert.Equal(t, tt.expected, result)
-		})
-	}
 }
 
 func TestOutputResult(t *testing.T) {
@@ -707,339 +193,4 @@ func TestTerraformOutputData(t *testing.T) {
 	assert.Len(t, data.ImportedResources, 1)
 	assert.Len(t, data.Outputs, 1)
 	assert.Contains(t, data.ChangedResult, "5 to add")
-}
-
-func TestExtractEventPrefix(t *testing.T) {
-	tests := []struct {
-		event    string
-		expected string
-	}{
-		{"before.terraform.plan", "before"},
-		{"after.terraform.plan", "after"},
-		{"after.helmfile.diff", "after"},
-		{"single", "single"},
-		{"", ""},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.event, func(t *testing.T) {
-			result := extractEventPrefix(tt.event)
-			assert.Equal(t, tt.expected, result)
-		})
-	}
-}
-
-func TestGetPluginAndBinding_EmptyEvent(t *testing.T) {
-	backup := testSaveAndClearRegistry()
-	defer testRestoreRegistry(backup)
-
-	ClearPlugins()
-	sp := &stubPlugin{
-		componentType: "terraform",
-		bindings: []plugin.HookBinding{
-			{Event: "after.terraform.plan", Handler: func(_ *plugin.HookContext) error { return nil }},
-		},
-	}
-	require.NoError(t, RegisterPlugin(sp))
-
-	// Empty event should return nil.
-	pl, binding := getPluginAndBinding(ExecuteOptions{Event: ""})
-	assert.Nil(t, pl)
-	assert.Nil(t, binding)
-}
-
-func TestGetPluginAndBinding_UnregisteredComponentType(t *testing.T) {
-	backup := testSaveAndClearRegistry()
-	defer testRestoreRegistry(backup)
-	ClearPlugins()
-
-	// No plugins registered.
-	pl, binding := getPluginAndBinding(ExecuteOptions{Event: "after.helmfile.diff"})
-	assert.Nil(t, pl)
-	assert.Nil(t, binding)
-}
-
-func TestGetPluginAndBinding_EventNotHandled(t *testing.T) {
-	backup := testSaveAndClearRegistry()
-	defer testRestoreRegistry(backup)
-
-	ClearPlugins()
-	sp := &stubPlugin{
-		componentType: "terraform",
-		bindings: []plugin.HookBinding{
-			{Event: "after.terraform.plan", Handler: func(_ *plugin.HookContext) error { return nil }},
-		},
-	}
-	require.NoError(t, RegisterPlugin(sp))
-
-	// Event exists for plugin type but specific event not handled.
-	pl, binding := getPluginAndBinding(ExecuteOptions{Event: "after.terraform.init"})
-	assert.Nil(t, pl)
-	assert.Nil(t, binding)
-}
-
-func TestGetPluginAndBinding_ComponentTypeOverride(t *testing.T) {
-	backup := testSaveAndClearRegistry()
-	defer testRestoreRegistry(backup)
-
-	ClearPlugins()
-	sp := &stubPlugin{
-		componentType: "terraform",
-		bindings: []plugin.HookBinding{
-			{Event: "after.terraform.plan", Handler: func(_ *plugin.HookContext) error { return nil }},
-		},
-	}
-	require.NoError(t, RegisterPlugin(sp))
-
-	// Use ComponentType override instead of extracting from event.
-	pl, binding := getPluginAndBinding(ExecuteOptions{
-		Event:         "after.terraform.plan",
-		ComponentType: "terraform",
-	})
-	assert.NotNil(t, pl)
-	assert.NotNil(t, binding)
-}
-
-func TestExecute_NilHandler(t *testing.T) {
-	backup := testSaveAndClearRegistry()
-	defer testRestoreRegistry(backup)
-
-	Register(&mockProvider{name: "generic", detected: false})
-
-	ClearPlugins()
-	sp := &stubPlugin{
-		componentType: "terraform",
-		bindings: []plugin.HookBinding{
-			{Event: "after.terraform.plan", Handler: nil}, // No handler.
-		},
-	}
-	require.NoError(t, RegisterPlugin(sp))
-
-	err := Execute(ExecuteOptions{
-		Event:       "after.terraform.plan",
-		ForceCIMode: true,
-		Info:        &schema.ConfigAndStacksInfo{Stack: "dev", ComponentFromArg: "vpc"},
-	})
-	assert.NoError(t, err)
-}
-
-func TestExecute_HandlerError_DoesNotPropagateToExecute(t *testing.T) {
-	backup := testSaveAndClearRegistry()
-	defer testRestoreRegistry(backup)
-
-	Register(&mockProvider{name: "generic", detected: false})
-
-	ClearPlugins()
-	sp := &stubPlugin{
-		componentType: "terraform",
-		bindings: []plugin.HookBinding{
-			{
-				Event: "after.terraform.plan",
-				Handler: func(_ *plugin.HookContext) error {
-					return fmt.Errorf("handler failed")
-				},
-			},
-		},
-	}
-	require.NoError(t, RegisterPlugin(sp))
-
-	// Handler errors are logged as warnings, not propagated.
-	err := Execute(ExecuteOptions{
-		Event:       "after.terraform.plan",
-		ForceCIMode: true,
-		AtmosConfig: &schema.AtmosConfiguration{},
-		Info:        &schema.ConfigAndStacksInfo{Stack: "dev", ComponentFromArg: "vpc"},
-	})
-	assert.NoError(t, err)
-}
-
-func TestDetectStoreFromEnv(t *testing.T) {
-	t.Run("no env vars returns nil", func(t *testing.T) {
-		// Clear relevant env vars.
-		t.Setenv("ATMOS_PLANFILE_BUCKET", "")
-		t.Setenv("GITHUB_ACTIONS", "")
-
-		result := detectStoreFromEnv()
-		assert.Nil(t, result)
-	})
-
-	t.Run("S3 bucket configured", func(t *testing.T) {
-		t.Setenv("ATMOS_PLANFILE_BUCKET", "my-bucket")
-		t.Setenv("ATMOS_PLANFILE_PREFIX", "plans/")
-		t.Setenv("AWS_REGION", "us-east-1")
-		t.Setenv("GITHUB_ACTIONS", "")
-
-		result := detectStoreFromEnv()
-		require.NotNil(t, result)
-		assert.Equal(t, "aws/s3", result.Type)
-		assert.Equal(t, "my-bucket", result.Options["bucket"])
-		assert.Equal(t, "plans/", result.Options["prefix"])
-		assert.Equal(t, "us-east-1", result.Options["region"])
-	})
-
-	t.Run("GitHub Actions detected", func(t *testing.T) {
-		t.Setenv("ATMOS_PLANFILE_BUCKET", "")
-		t.Setenv("GITHUB_ACTIONS", "true")
-
-		result := detectStoreFromEnv()
-		require.NotNil(t, result)
-		assert.Equal(t, "github/artifacts", result.Type)
-	})
-
-	t.Run("S3 takes precedence over GitHub", func(t *testing.T) {
-		t.Setenv("ATMOS_PLANFILE_BUCKET", "bucket")
-		t.Setenv("GITHUB_ACTIONS", "true")
-
-		result := detectStoreFromEnv()
-		require.NotNil(t, result)
-		assert.Equal(t, "aws/s3", result.Type)
-	})
-}
-
-func TestCreatePlanfileStore(t *testing.T) {
-	t.Run("defaults to local store", func(t *testing.T) {
-		t.Setenv("ATMOS_PLANFILE_BUCKET", "")
-		t.Setenv("GITHUB_ACTIONS", "")
-
-		store, err := createPlanfileStore(ExecuteOptions{
-			AtmosConfig: &schema.AtmosConfiguration{},
-		})
-		require.NoError(t, err)
-		assert.NotNil(t, store)
-		assert.Equal(t, "local/dir", store.Name())
-	})
-
-	t.Run("nil config defaults to local store", func(t *testing.T) {
-		t.Setenv("ATMOS_PLANFILE_BUCKET", "")
-		t.Setenv("GITHUB_ACTIONS", "")
-
-		store, err := createPlanfileStore(ExecuteOptions{
-			AtmosConfig: nil,
-		})
-		require.NoError(t, err)
-		assert.NotNil(t, store)
-		assert.Equal(t, "local/dir", store.Name())
-	})
-}
-
-func TestBuildHookContext(t *testing.T) {
-	mp := &mockProvider{name: "test", detected: true}
-
-	opts := ExecuteOptions{
-		Event:       "after.terraform.plan",
-		AtmosConfig: &schema.AtmosConfiguration{},
-		Info: &schema.ConfigAndStacksInfo{
-			Stack:            "dev",
-			ComponentFromArg: "vpc",
-		},
-		Output:       "some output",
-		CommandError: fmt.Errorf("some error"),
-	}
-
-	ctx := buildHookContext(opts, mp)
-
-	assert.Equal(t, "after.terraform.plan", ctx.Event)
-	assert.Equal(t, "plan", ctx.Command)
-	assert.Equal(t, "after", ctx.EventPrefix)
-	assert.Equal(t, opts.AtmosConfig, ctx.Config)
-	assert.Equal(t, opts.Info, ctx.Info)
-	assert.Equal(t, "some output", ctx.Output)
-	assert.Equal(t, opts.CommandError, ctx.CommandError)
-	assert.Equal(t, mp, ctx.Provider)
-	assert.NotNil(t, ctx.TemplateLoader)
-	assert.NotNil(t, ctx.CreatePlanfileStore)
-}
-
-func TestExecute_SummaryContentFlowsToOutput(t *testing.T) {
-	// Verify that the rendered summary markdown from the summary handler
-	// is available as a "summary" output variable.
-
-	// Create a temp directory with a template file.
-	tmpDir := t.TempDir()
-	terraformDir := filepath.Join(tmpDir, "terraform")
-	require.NoError(t, os.MkdirAll(terraformDir, 0o755))
-	templateContent := "## Plan for `{{.Component}}` in `{{.Stack}}`\n"
-	require.NoError(t, os.WriteFile(filepath.Join(terraformDir, "plan.md"), []byte(templateContent), 0o644))
-
-	// Set up provider registry with a full capturing provider.
-	backup := testSaveAndClearRegistry()
-	defer testRestoreRegistry(backup)
-
-	fcp := newFullCapturingProvider("generic", false)
-	Register(fcp)
-
-	// Set up plugin with a handler that renders summary and writes outputs.
-	ClearPlugins()
-	sp := &stubPlugin{
-		componentType: "terraform",
-		bindings: []plugin.HookBinding{
-			{
-				Event: "after.terraform.plan",
-				Handler: func(ctx *plugin.HookContext) error {
-					writer := ctx.Provider.OutputWriter()
-					if writer == nil {
-						return nil
-					}
-
-					// Render summary template.
-					tmplCtx := map[string]string{"Component": ctx.Info.ComponentFromArg, "Stack": ctx.Info.Stack}
-					rendered, err := ctx.TemplateLoader.LoadAndRender("terraform", "plan", nil, tmplCtx)
-					if err != nil {
-						return err
-					}
-
-					// Write summary.
-					if err := writer.WriteSummary(rendered); err != nil {
-						return err
-					}
-
-					// Write outputs.
-					vars := map[string]string{
-						"has_changes": "true",
-						"summary":     rendered,
-						"stack":       ctx.Info.Stack,
-						"component":   ctx.Info.ComponentFromArg,
-						"command":     ctx.Command,
-					}
-					for key, value := range vars {
-						if err := writer.WriteOutput(key, value); err != nil {
-							return err
-						}
-					}
-
-					return nil
-				},
-			},
-		},
-	}
-	require.NoError(t, RegisterPlugin(sp))
-
-	// Execute with CI forced, using template base_path pointing to temp dir.
-	err := Execute(ExecuteOptions{
-		Event:       "after.terraform.plan",
-		ForceCIMode: true,
-		AtmosConfig: &schema.AtmosConfiguration{
-			CI: schema.CIConfig{
-				Templates: schema.CITemplatesConfig{
-					BasePath: tmpDir,
-				},
-			},
-		},
-		Info: &schema.ConfigAndStacksInfo{
-			Stack:            "dev-us-east-1",
-			ComponentFromArg: "vpc",
-		},
-		Output: "some terraform output",
-	})
-	require.NoError(t, err)
-
-	// Verify summary was written.
-	require.Len(t, fcp.writer.summaries, 1, "WriteSummary should have been called once")
-	assert.Contains(t, fcp.writer.summaries[0], "vpc")
-	assert.Contains(t, fcp.writer.summaries[0], "dev")
-
-	// Verify the "summary" output variable contains the rendered summary markdown.
-	summaryOutput, ok := fcp.writer.outputs["summary"]
-	assert.True(t, ok, "output variables should include 'summary'")
-	assert.Equal(t, fcp.writer.summaries[0], summaryOutput, "summary output should match rendered summary")
 }
